+++ mobile-playground/metro.config.js
@@ -1,0 +1,17 @@
+const path = require('path');
+
+module.exports = {
+  projectRoot: __dirname,
+  watchFolders: [
+    path.resolve(__dirname, '..'), // cds
+    path.resolve(__dirname, '../../../../node_modules'),
+  ],
+  transformer: {
+    getTransformOptions: async () => ({
+      transform: {
+        experimentalImportSupport: false,
+        inlineRequires: false,
+      },
+    }),
+  },
+};