+++ mobile-playground/src/TextHeadline.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { TextHeadline } from '@cds/mobile';
+
+import TextExamplesScreen from './internal/TextExamples';
+
+// eslint-disable-next-line react/display-name
+export default () => <TextExamplesScreen component={TextHeadline} />;