+++ mobile-playground/src/TextLabel1.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { TextLabel1 } from '@cds/mobile';
+
+import TextExamplesScreen from './internal/TextExamples';
+
+// eslint-disable-next-line react/display-name
+export default () => <TextExamplesScreen component={TextLabel1} />;