+++ mobile-playground/src/TextLegal.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { TextLegal } from '@cds/mobile';
+
+import TextExamplesScreen from './internal/TextExamples';
+
+// eslint-disable-next-line react/display-name
+export default () => <TextExamplesScreen component={TextLegal} />;