+++ mobile-playground/src/TextTitle3.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { TextTitle3 } from '@cds/mobile';
+
+import TextExamplesScreen from './internal/TextExamples';
+
+// eslint-disable-next-line react/display-name
+export default () => <TextExamplesScreen component={TextTitle3} />;