+++ mobile-playground/src/internal/Example.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import { TextTitle3 } from '@cds/mobile';
+import { usePalette } from '@cds/theme/native';
+import { View } from 'react-native';
+
+interface ExampleProps {
+  children: React.ReactNode;
+  title?: string;
+}
+
+const Example = ({ children, title }: ExampleProps) => {
+  const palette = usePalette();
+  const items = React.Children.toArray(children);
+
+  return (
+    <View
+      style={{
+        borderBottomWidth: 1,
+        borderBottomColor: palette.stroke,
+        padding: 16,
+        paddingBottom: 24,
+      }}
+    >
+      {title && <TextTitle3>{title}</TextTitle3>}
+
+      {items.map((item, index) => (
+        <View key={index} style={{ paddingTop: 12 }}>
+          {item}
+        </View>
+      ))}
+    </View>
+  );
+};
+
+export default Example;